+++ packages/terra-button/src/index.js
@@ -1,0 +1,9 @@
+import Button from './Button';
+import IconButton from './variants/_IconButton';
+import { ButtonTypes, ButtonVariants, IconTypes } from './_constants';
+
+export default Button;
+
+export {
+  IconButton, ButtonTypes, ButtonVariants, IconTypes,
+};