+++ packages/terra-core-docs/src/terra-dev-site/test/tag/MultipleTags.test.jsx
@@ -1,6 +1,6 @@
 /* eslint-disable no-alert */
 import React from 'react';
-import IconBookmark from 'terra-icon/lib/icon/IconTile';
+import IconBookmark from 'terra-icon/lib/icon/decorative/IconTile';
 import Tag from 'terra-tag';
 
 const icon = <IconBookmark />;