+++ packages/terra-search-field/src/SearchField.jsx
@@ -3,15 +3,13 @@ import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import classNamesBind from 'classnames/bind';
 import ThemeContext from 'terra-theme-context';
-import Button from 'terra-button';
+import { IconButton } from 'terra-button';
 import * as KeyCode from 'keycode-js';
 import IconSearch from 'terra-icon/lib/icon/IconSearch';
 import { injectIntl } from 'react-intl';
 import styles from './SearchField.module.scss';
 
 const cx = classNamesBind.bind(styles);
-
-const Icon = <IconSearch />;
 
 const propTypes = {
   /**
@@ -270,13 +268,14 @@ class SearchField extends React.Component {
 
     const clearButton = this.searchText && !isDisabled
       ? (
-        <Button
+        <IconButton
           data-terra-search-field-button="Clear"
           className={cx('clear')}
           onClick={this.handleClear}
           text={clearText}
           variant="utility"
           icon={clearIcon}
+          iconType={IconButton.Opts.IconTypes.INFORMATIVE}
           isIconOnly
         />
       )
@@ -300,13 +299,14 @@ class SearchField extends React.Component {
           />
           {clearButton}
         </div>
-        <Button
+        <IconButton
           data-terra-search-field-button="Search"
           className={cx('button')}
           text={buttonText}
           onClick={this.handleSearch}
           isDisabled={isDisabled}
-          icon={Icon}
+          icon={<IconSearch a11yLabel={intl.formatMessage({ id: 'Terra.searchField.search' })} />}
+          iconType={IconButton.Opts.IconTypes.INFORMATIVE}
           isIconOnly
           isCompact
           refCallback={(ref) => { this.searchBtnRef.current = ref; }}