+++ packages/terra-dialog/CHANGELOG.md
@@ -5,6 +5,11 @@
 * Changed
   * Replaced `terra-button` with `IconButton`
 
+## 2.63.6 - (July 5, 2022)
+
+* Changed
+  * Minor dependency version bump
+
 ## 2.63.5 - (May 3, 2022)
 
 * Changed